Fixes bug where menubar would not close
When the bottom of the page is clicked, the menubar would not close.
It turns out that the div containing the whole app does not stretch to fit
all the space of the screen.

This commit sets the whole app's height to stretch to the space of the screen,
therefore fixing this bug.